Move buildXTrain to Model class
